removing duplication code
