+++ centreon/packages/js-config/webpack/patch/devServer.js
@@ -1,0 +1,37 @@
+const os = require('os');
+
+const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
+
+const devServerPort = 9090;
+
+const interfaces = os.networkInterfaces();
+const externalInterface = Object.keys(interfaces).find(
+  (interfaceName) =>
+    !interfaceName.includes('docker') &&
+    interfaces[interfaceName][0].family === 'IPv4' &&
+    interfaces[interfaceName][0].internal === false &&
+    !process.env.IS_STATIC_PORT_FORWARDED
+);
+
+const devServerAddress = externalInterface
+  ? interfaces[externalInterface][0].address
+  : 'localhost';
+
+const publicPath = `http://${devServerAddress}:${devServerPort}/static/`;
+
+const isServeMode = process.env.WEBPACK_ENV === 'serve';
+const isDevelopmentMode = process.env.WEBPACK_ENV === 'development';
+
+const devServerPlugins = isServeMode ? [new ReactRefreshWebpackPlugin()] : [];
+
+module.exports = {
+  devServer: {
+    compress: true,
+    hot: true,
+    port: devServerPort
+  },
+  devServerPlugins,
+  isDevelopmentMode,
+  isServeMode,
+  publicPath
+};
